Add files via upload
Anyone want to chime in on this DAPP I'm developing? The concept is a platform that resembles a  global group chat that allows any idea, you originally created to be fully owned by you secure on the blockchain to not be changed or manipulated, with this Dapp you can collaborate, invest, sell and buy ideas globally granting the people of the world the ability to work together on a decentralized platform to create positive change in the world  

For example, someone in Florida can upload their idea discussing the lack of female trees and how to increase the growth and protection of female trees on this application and an agriculturist from south Africa with the resources or platform has the choice to buy, invest or collaborate on the concept, someone is currently to helping fight that problem now, but no one knows about it.